+++ systest/21million/live/run_test.go
@@ -19,30 +19,10 @@
 package bulk
 
 import (
-	"context"
-	"fmt"
-	"io"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
-	"strings"
-	"time"
-
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dgraph-io/dgraph/chunker"
-	"github.com/dgraph-io/dgraph/dgraphtest"
+	"github.com/dgraph-io/dgraph/systest/21million/common"
 )
-
-var datafiles = map[string]string{
-	"1million-noindex.schema": "https://github.com/dgraph-io/benchmarks/blob/master/data/1million-noindex.schema?raw=true",
-	"1million.schema":         "https://github.com/dgraph-io/benchmarks/blob/master/data/1million.schema?raw=true",
-	"1million.rdf.gz":         "https://github.com/dgraph-io/benchmarks/blob/master/data/1million.rdf.gz?raw=true",
-	"21million.schema":        "https://github.com/dgraph-io/benchmarks/blob/master/data/21million.schema?raw=true",
-	"21million.rdf.gz":        "https://github.com/dgraph-io/benchmarks/blob/master/data/21million.rdf.gz?raw=true",
-}
 
 func (lsuite *LiveTestSuite) TestQueriesFor21Million() {
 	t := lsuite.T()
@@ -51,67 +31,5 @@ func (lsuite *LiveTestSuite) TestQueriesFor21Million() {
 	// Upgrade
 	lsuite.Upgrade()
 
-	dg, cleanup, err := lsuite.dc.Client()
-	defer cleanup()
-	require.NoError(t, err)
-
-	_, thisFile, _, _ := runtime.Caller(0)
-	queryDir := filepath.Join(filepath.Dir(thisFile), "../queries")
-
-	// For this test we DON'T want to start with an empty database.
-	files, err := os.ReadDir(queryDir)
-	if err != nil {
-		t.Fatalf("Error reading directory: %s", err.Error())
-	}
-
-	for _, file := range files {
-		if !strings.HasPrefix(file.Name(), "query-") {
-			continue
-		}
-		lsuite.Run(file.Name(), func() {
-			filename := filepath.Join(queryDir, file.Name())
-			reader, cleanup := chunker.FileReader(filename, nil)
-			bytes, err := io.ReadAll(reader)
-			if err != nil {
-				t.Fatalf("Error reading file: %s", err.Error())
-			}
-			contents := string(bytes[:])
-			cleanup()
-
-			// The test query and expected result are separated by a delimiter.
-			bodies := strings.SplitN(contents, "\n---\n", 2)
-			// Dgraph can get into unhealthy state sometime. So, add retry for every query.
-			for retry := 0; retry < 3; retry++ {
-				// If a query takes too long to run, it probably means dgraph is stuck and there's
-				// no point in waiting longer or trying more tests.
-				ctx, cancel := context.WithTimeout(context.Background(), 10*time.Minute)
-				resp, err := dg.NewTxn().Query(ctx, bodies[0])
-				cancel()
-
-				if retry < 2 && (err != nil || ctx.Err() == context.DeadlineExceeded) {
-					continue
-				}
-
-				if ctx.Err() == context.DeadlineExceeded {
-					t.Fatal("aborting test due to query timeout")
-				}
-
-				t.Logf("running %s", file.Name())
-
-				dgraphtest.CompareJSON(bodies[1], string(resp.GetJson()))
-			}
-		})
-	}
-}
-
-func downloadDataFiles(testDir string) error {
-	for fname, link := range datafiles {
-		cmd := exec.Command("wget", "-O", fname, link)
-		cmd.Dir = testDir
-
-		if out, err := cmd.CombinedOutput(); err != nil {
-			return errors.Wrapf(err, fmt.Sprintf("error downloading a file: %s", string(out)))
-		}
-	}
-	return nil
+	require.NoError(t, common.QueriesFor21Million(lsuite.T(), lsuite.dc))
 }
